# apps/itsm/classic/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ITSM Classic",
    "version": "0.1.0"
  },
  "paths": {
    "/api/incidents": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Incident",
        "operationId": "createIncident",
        "parameters": [
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "listIncidents",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Incident"
                  },
                  "title": "Response Listincidents"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "getIncident",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/assign": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Assign Incident",
        "operationId": "assignIncident",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/request-info": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Request Info",
        "operationId": "requestInfo",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/provide-info": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Provide Info",
        "operationId": "provideInfo",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/escalate": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Escalate Incident",
        "operationId": "escalateIncident",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/resolve": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Resolve Incident",
        "operationId": "resolveIncident",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/close": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Close Incident",
        "operationId": "closeIncident",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/reopen": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Reopen Incident",
        "operationId": "reopenIncident",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/comments": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Add Comment",
        "operationId": "addIncidentComment",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/priority": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Update Priority",
        "operationId": "updateIncidentPriority",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePriorityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/link/ci": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Link Ci",
        "operationId": "linkIncidentCi",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCiRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{iid}/link/change": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Link Change",
        "operationId": "linkIncidentChange",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/changes/{cid}/link/incident": {
      "post": {
        "tags": [
          "changes"
        ],
        "summary": "Link Change Incident",
        "operationId": "linkChangeIncident",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "itsm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Itsm Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddCommentRequest": {
        "properties": {
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "AddCommentRequest"
      },
      "AssignRequest": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          }
        },
        "type": "object",
        "title": "AssignRequest"
      },
...
      "CreateIncidentRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "priority": {
            "type": "string",
            "enum": [
              "P1",
              "P2",
              "P3",
              "P4"
            ],
            "title": "Priority"
          },
          "category": {
            "type": "string",
            "enum": [
              "hardware",
              "software",
              "network",
              "security",
              "performance"
            ],
            "title": "Category"
          },
          "affected_ci": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Ci"
          }
        },
        "type": "object",
        "required": [
          "title",
          "priority",
          "category"
        ],
        "title": "CreateIncidentRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Incident": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "P1",
              "P2",
              "P3",
              "P4"
            ],
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_progress",
              "pending_user",
              "resolved",
              "closed",
              "escalated"
            ],
            "title": "Status"
          },
          "category": {
            "type": "string",
            "enum": [
              "hardware",
              "software",
              "network",
              "security",
              "performance"
            ],
            "title": "Category"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "requester": {
            "type": "string",
            "title": "Requester"
          },
          "affected_ci": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Ci"
          },
          "linked_cis": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Cis"
          },
          "linked_changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Changes"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "type": "array",
            "title": "Comments"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "sla_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            },
            "type": "array",
            "title": "History"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "status",
          "category",
          "requester",
          "created_at",
          "updated_at"
        ],
        "title": "Incident"
      },
      "LinkChangeRequest": {
        "properties": {
          "change_id": {
            "type": "string",
            "title": "Change Id"
          }
        },
        "type": "object",
        "required": [
          "change_id"
        ],
        "title": "LinkChangeRequest"
      },
      "LinkCiRequest": {
        "properties": {
          "ci_id": {
            "type": "string",
            "title": "Ci Id"
          }
        },
        "type": "object",
        "required": [
          "ci_id"
        ],
        "title": "LinkCiRequest"
      },
      "LinkIncidentRequest": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          }
        },
        "type": "object",
        "required": [
          "incident_id"
        ],
        "title": "LinkIncidentRequest"
      },
      "ResolveIncidentRequest": {
        "properties": {
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          }
        },
        "type": "object",
        "title": "ResolveIncidentRequest"
      },
...
      "UpdatePriorityRequest": {
        "properties": {
          "priority": {
            "type": "string",
            "enum": [
              "P1",
              "P2",
              "P3",
              "P4"
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "priority"
        ],
        "title": "UpdatePriorityRequest"
      },
...
    }
  }
}